inference/requirements.txt
numpy
torch
onnxruntime-gpu
transformers
influxdb-client
flask